+++ Changelog.md
@@ -3,6 +3,7 @@
 - prevent mounting selection canvas when viewer mode
 - prevent drawing selection when not needed
 - Click on toolbar tool group always enable a tool
+- fix input and select options not being impacted by theme
 
 # V.0.4.0